Move verification logic in model sort function to enterprise folder
