+++ enterprise/frontend/src/metabase-enterprise/content_verification/index.ts
@@ -2,7 +2,11 @@ import { PLUGIN_CONTENT_VERIFICATION } from "metabase/plugins";
 import { hasPremiumFeature } from "metabase-enterprise/settings";
 import { VerifiedFilter } from "./VerifiedFilter";
 import { BrowseFilterControls } from "./BrowseFilterControls";
-import { browseFilters, sortCollectionsForBrowseModels } from "./utils";
+import {
+  browseFilters,
+  sortCollectionsForBrowseModels,
+  sortModelsByVerification,
+} from "./utils";
 
 if (hasPremiumFeature("content_verification")) {
   Object.assign(PLUGIN_CONTENT_VERIFICATION, {
@@ -10,5 +14,6 @@ if (hasPremiumFeature("content_verification")) {
     BrowseFilterControls,
     browseFilters,
     sortCollectionsForBrowseModels,
+    sortModelsByVerification,
   });
 }
